+++ sdk/core/Azure.Core/src/JsonPatchOperationKind.cs
@@ -7,7 +7,7 @@ namespace Azure.Core.JsonPatch
     {
         private readonly string _operation;
 
-        public JsonPatchOperationKind(string operation)
+        private JsonPatchOperationKind(string operation)
         {
             _operation = operation;
         }