+++ sdk/core/Azure.Core.Experimental/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.1.0-preview.7 (Unreleased)
 
+### Breaking Changes
+- `JsonPatchDocument` type moved to `Azure.Core` package.
 
 ## 0.1.0-preview.6 (2020-10-06)